adjusted json file path
For testing the program it was executed not directly from the build directory but from its parent. When starting the program in the build the config file specified by a relative path, couldn't be found. The program now expects the config file to be located in the project folder, and needs to be executed from the build folder.